RaceProcessor 가 진행에 필요한 메소드를 제공하도록 변경
