+++ src/test/java/RacingCarApplicationTest.java
@@ -1,0 +1,10 @@
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class RacingCarApplicationTest {
+
+    @Test
+    void main() {
+    }
+}
